Fixed a slight bug on line 15 of fuel/application/config/config.php
I'm on a windows box with Apache 2.2. and php 5.2 installed.

Apache is not installed on the default port 80 but on a different port. But after I instsll FUEL-CMS the admin login screen was always opened with the default port 80. Line 15, i discovered, was were the problem lay. 

I suspect that it might be windows only issue tho.