Related to fuel issue #246
The fuel_navigation_model validation is allowing empty locations to be created and modified.   This can lead to strange behavior down the line when you create another empty location with the same parent.

Looking at the code it seems the intent is to never have a empty location.   So I have modified is_editable_location and is_new_location to return FALSE if the location is empty.